Update #namUSGrab.py#
This code grabs all of the case file images from a give case and iterates over the pages which have status codes of 200 in order to eliminate any bad pages. It then compiles a case file with link to the case url as well as the urls of each of that case's images.